requirements.txt
python-telegram-bot==21.6
openai==1.58.1
google-genai==1.5.0
anthropic==0.42.0
pyyaml==6.0.2
python-dotenv==1.0.1
